la logica del endpoint la coloque en una carpeta aparte llamado controllers
